requirements.txt
openbb
requests
yfinance
pandas
streamlit
google-generativeai
sqlalchemy
python-dotenv
apscheduler